pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fuck-u-code"
version = "1.0.0"
description = "代码质量分析工具 - Python版本"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "fuck-u-code Team", email = "team@fuck-u-code.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0.0",
    "rich>=12.0.0",
    "pyyaml>=6.0",
    "esprima>=4.0.1",
    "javalang>=0.13.0",
    "pycparser>=2.20",
    "typing-extensions>=4.0.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "isort>=5.10.0",
]

[project.scripts]
fuck-u-code = "fuck_u_code.cli.main:main"
fuc = "fuck_u_code.cli.main:main"

[project.urls]
Homepage = "https://github.com/fuck-u-code/fuck-u-code-python"
Repository = "https://github.com/fuck-u-code/fuck-u-code-python"
Documentation = "https://fuck-u-code-python.readthedocs.io"
"Bug Tracker" = "https://github.com/fuck-u-code/fuck-u-code-python/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fuck_u_code = ["i18n/*.json", "templates/*.md"]

# 工具配置
[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]